refactor(courses): consolidate course dashboard into route
Moves the `CourseDashboard` component and its related sub-components
from `pages/shared/courses/course-dashboard.tsx` directly into
`routes/(protected)/courses/index.tsx`.
This change eliminates the need for the `RoleComponentManager`
wrapper and integrates course-related mock data within the route.
Also removes the `courseCode` prop from `ModulePanel` in module
routes, as it can be derived from the module object.
Adds a loader for academic terms and updates the
`CourseCardActionButton` to include a 'Start Meeting' option for
the mentor role.
