+++ frontend/src/features/courses/modules/module-review-step.tsx
@@ -86,12 +86,7 @@ const ModuleReviewStep = ({ module }: ReviewStepProps) => {
         </Tabs.List>
 
         <Tabs.Panel value="preview" pt="md">
-          <ModulePanel
-            allExpanded={true}
-            isPreview={true}
-            module={module}
-            courseCode={module.courseCode}
-          />
+          <ModulePanel allExpanded={true} isPreview={true} module={module} />
         </Tabs.Panel>
 
         <Tabs.Panel value="summary" pt="md">
