+++ frontend/src/routes/(protected)/courses/index.tsx
@@ -1,50 +1,569 @@
-import { createFileRoute } from '@tanstack/react-router'
-import RoleComponentManager from '@/components/role-component-manager.tsx'
+import { createFileRoute, Link } from '@tanstack/react-router'
 import { useAuth } from '@/features/auth/auth.hook.ts'
-import type { Course, EnrolledAcademicTerm } from '@/features/courses/types'
-import CourseDashboard from '@/pages/shared/courses/course-dashboard.tsx'
+import type {
+  AcademicProgram,
+  ClassMeeting,
+  Course,
+  EnrolledAcademicTerm,
+} from '@/features/courses/types'
+import { useState } from 'react'
+import {
+  Box,
+  Button,
+  Card,
+  Container,
+  Divider,
+  Flex,
+  Grid,
+  Group,
+  Image,
+  Progress,
+  RingProgress,
+  Select,
+  Stack,
+  Text,
+  Title,
+  useMantineTheme,
+} from '@mantine/core'
+import SearchComponent from '@/components/search-component.tsx'
+import {
+  IconDeviceDesktop,
+  IconLayoutGridFilled,
+  IconList,
+  IconVideo,
+} from '@tabler/icons-react'
+import CourseTasksSummary from '@/features/courses/course-task-summary.tsx'
+import { useCurrentMeeting } from '@/features/courses/hooks/useCurrentMeeting.ts'
+import CourseDashboardQuickActions from '@/features/courses/dashboard/course-dashboard-quick-actions.tsx'
 
 export const Route = createFileRoute('/(protected)/courses/')({
   component: RouteComponent,
+  loader: ({ context }) => {
+    return mockTerms //TODO: replace this with actual fetch
+  },
 })
 
+const mockTerms: EnrolledAcademicTerm[] = [
+  {
+    termId: 'termId1',
+    schoolYear: 'SY 2024-2025',
+    term: 'Term 1',
+    isCurrent: false,
+  },
+  {
+    termId: 'termId2',
+    schoolYear: 'SY 2024-2025',
+    term: 'Term 2',
+    isCurrent: false,
+  },
+  {
+    termId: 'termId3',
+    schoolYear: 'SY 2024-2025',
+    term: 'Term 3',
+    isCurrent: false,
+  },
+  {
+    termId: 'termId4',
+    schoolYear: 'SY 2025-2026',
+    term: 'Term 1',
+    isCurrent: true,
+  },
+]
+
+// TODO: Consider adding program and/or department and major to the course data
+// TODO: Course types might also be necessary such as 'General Education', 'Specialization', etc.
+const MockCourseData: Course[] = [
+  {
+    courseDetails: {
+      courseName: 'Web Technology Applications',
+      courseCode: 'MO-IT200',
+    },
+    courseProgress: 0.5,
+    section: {
+      sectionName: 'A2101',
+      sectionSchedule: {
+        day: 'MWF',
+        time: '10:00 - 11:00 AM',
+      },
+      classMeetings: [
+        {
+          startTimeStamp: '2023-08-20T10:00',
+          endTimeStamp: '2023-08-20T11:00',
+          meetingLink: 'https://zoom.us',
+        },
+      ],
+    },
+
+    activities: [],
+  },
+  {
+    courseDetails: {
+      courseName: 'Data Structures and Algorithms',
+      courseCode: 'MO-IT351',
+    },
+    courseProgress: 0.5,
+    section: {
+      sectionName: 'A2101',
+      sectionSchedule: {
+        day: 'TTHS',
+        time: '10:00 - 11:00 AM',
+      },
+      classMeetings: [
+        {
+          startTimeStamp: '2023-08-20T10:00',
+          endTimeStamp: '2023-08-20T11:00',
+          meetingLink: 'https://zoom.us',
+        },
+      ],
+    },
+    activities: [],
+  },
+
+  {
+    courseDetails: {
+      courseName: 'Capstone 1',
+      courseCode: 'MO-IT400',
+    },
+    courseProgress: 0.5,
+    section: {
+      sectionName: 'A2101',
+      sectionSchedule: {
+        day: 'MWF',
+        time: '8:00 - 9:00 AM',
+      },
+      classMeetings: [
+        {
+          startTimeStamp: '2023-08-20T15:45',
+          endTimeStamp: '2023-08-20T22:00',
+          meetingLink: 'https://zoom.us',
+        },
+      ],
+    },
+    activities: [],
+  },
+  {
+    courseDetails: {
+      courseName: 'Capstone 2',
+      courseCode: 'MO-IT500',
+    },
+    courseProgress: 0.5,
+    section: {
+      sectionName: 'A2101',
+      sectionSchedule: {
+        day: 'TTHS',
+        time: '8:00 - 9:00 AM',
+      },
+      classMeetings: [
+        {
+          startTimeStamp: '2025-08-23T00:00',
+          endTimeStamp: '2025-08-29T23:59',
+          meetingLink: 'https://zoom.us',
+        },
+      ],
+    },
+    activities: [
+      {
+        activityName: 'Assignment 1',
+        dueTimestamp: '2025-08-20T23:59:59',
+      },
+      {
+        activityName: 'Assignment 2',
+        dueTimestamp: '2025-08-20T23:59:59',
+      },
+    ],
+  },
+]
+
+const mockAcademicPrograms: AcademicProgram[] = [
+  {
+    program: 'Bachelor of Science in Information Technology',
+    programCode: 'BSIT',
+    major: 'Software Development',
+    majorCode: 'SD',
+  },
+  {
+    program: 'Bachelor of Science in Computer Science',
+    programCode: 'BSCS',
+    major: 'Software Engineering',
+    majorCode: 'SE',
+  },
+  {
+    program: 'Bachelor of Science in Information Systems',
+    programCode: 'BSIS',
+    major: 'Information Systems',
+    majorCode: 'IS',
+  },
+]
+
 function RouteComponent() {
-  const { authUser } = useAuth('protected')
-  const enrolledTerms: EnrolledAcademicTerm[] = [
-    {
-      termId: 'termId1',
-      schoolYear: 'SY 2024-2025',
-      term: 'Term 1',
-      isCurrent: false,
-    } as EnrolledAcademicTerm,
-    {
-      termId: 'termId2',
-      schoolYear: 'SY 2024-2025',
-      term: 'Term 2',
-      isCurrent: false,
-    } as EnrolledAcademicTerm,
-    {
-      termId: 'termId3',
-      schoolYear: 'SY 2024-2025',
-      term: 'Term 3',
-      isCurrent: false,
-    } as EnrolledAcademicTerm,
-    {
-      termId: 'termId4',
-      schoolYear: 'SY 2025-2026',
-      term: 'Term 1',
-      isCurrent: true,
-    } as EnrolledAcademicTerm,
-  ]
+  const { authUser } = useAuth('protected') //TODO: use this later for fetching enrolled terms
+  const academicTerms = Route.useLoaderData() //TODO: replace with suspense query
+
+  const theme = useMantineTheme()
+  const [view, setView] = useState<'grid' | 'list'>('grid')
+
+  const formatTerm = (academicTerm: EnrolledAcademicTerm | undefined) => {
+    return academicTerm
+      ? `${academicTerm.schoolYear} - ${academicTerm.term}`
+      : 'N/A'
+  }
+
+  const currentTerm = formatTerm(academicTerms.find((term) => term.isCurrent))
+
+  //TODO: implement API call to get courses, get it by academic term Id
+  const coursesData = MockCourseData
+  const [filteredCourses, setFilteredCourses] = useState<Course[]>(coursesData)
+
+  const academicPrograms = mockAcademicPrograms
+  const [selectedAcademicProgram, setSelectedAcademicProgram] =
+    useState<AcademicProgram>(academicPrograms[0])
 
   return (
-    <RoleComponentManager
-      currentRole={authUser.role}
-      roleRender={{
-        student: <CourseDashboard academicTerms={enrolledTerms} />,
-        admin: <CourseDashboard academicTerms={enrolledTerms} />, //TODO: change to all terms
-        // mentor: <CoursesMentorPage />,
-      }}
+    <Container fluid m={0}>
+      <Stack gap={'lg'}>
+        {/* Page Hero */}
+        <Group justify="space-between" align="center">
+          <Title c={'dark.7'} variant="hero" order={2} fw={700}>
+            All Courses
+          </Title>
+        </Group>
+
+        <Divider />
+
+        <Grid>
+          <Grid.Col span={authUser.role === 'student' ? 9 : 12}>
+            <Stack gap={'md'} mr={'md'}>
+              {/*Filters*/}
+              <Group justify="space-between" align="start">
+                <SearchComponent //TODO: might need to replace this later if paging should be implemented
+                  data={coursesData}
+                  onFilter={setFilteredCourses}
+                  identifiers={[['courseDetails', 'courseName']]}
+                  placeholder={'Search courses'}
+                />
+                <Group gap={'md'} align="center" justify="flex-end">
+                  {authUser.role !== 'student' && (
+                    <Select
+                      data={academicPrograms.map(
+                        (program) => program.programCode,
+                      )}
+                      defaultValue={selectedAcademicProgram.programCode}
+                      allowDeselect={false}
+                      variant="default"
+                      radius={'md'}
+                      size={'sm'}
+                      w={'20%'}
+                    />
+                  )}
+                  <Select
+                    data={academicTerms.map(
+                      (term) =>
+                        `${term.isCurrent ? '(Current)' : ''} ${formatTerm(term)}`,
+                    )}
+                    defaultValue={`(Current) ${currentTerm}`}
+                    allowDeselect={false}
+                    variant="default"
+                    radius={'md'}
+                    size={'sm'}
+                  />
+                  <Button.Group>
+                    <Button
+                      variant="default"
+                      radius={'md'}
+                      bg={
+                        view === 'grid'
+                          ? theme.colors.gray[3]
+                          : theme.colors.gray[0]
+                      }
+                      size={'xs'}
+                      onClick={() => setView('grid')}
+                    >
+                      <IconLayoutGridFilled
+                        size="75%"
+                        color={view === 'grid' ? 'black' : theme.colors.dark[2]}
+                      />
+                    </Button>
+                    <Button
+                      variant="default"
+                      radius={'md'}
+                      bg={
+                        view === 'list'
+                          ? theme.colors.gray[3]
+                          : theme.colors.gray[0]
+                      }
+                      size={'xs'}
+                      onClick={() => setView('list')}
+                    >
+                      <IconList
+                        size="75%"
+                        color={view === 'list' ? 'black' : theme.colors.dark[2]}
+                      />
+                    </Button>
+                  </Button.Group>
+                </Group>
+              </Group>
+              {/*Courses*/}
+              <Flex gap={'md'} wrap={'wrap'}>
+                {filteredCourses.map((course, index) => (
+                  <CourseItem key={index} course={course} variant={view} />
+                ))}
+              </Flex>
+            </Stack>
+          </Grid.Col>
+          {/*Upcoming Tasks*/}
+          {authUser.role === 'student' && (
+            <Grid.Col span={3}>
+              <CourseTasksSummary courses={filteredCourses} />
+            </Grid.Col>
+          )}
+        </Grid>
+      </Stack>
+    </Container>
+  )
+}
+
+const CourseItem = ({
+  course,
+  variant,
+}: {
+  course: Course
+  variant: 'grid' | 'list'
+}) => {
+  const url = `/courses/${course.courseDetails.courseCode}`
+  return variant === 'grid' ? (
+    <CourseCard
+      courseDetails={course.courseDetails}
+      courseProgress={course.courseProgress}
+      section={course.section}
+      url={url}
     />
+  ) : (
+    <CourseListRow
+      courseDetails={course.courseDetails}
+      courseProgress={course.courseProgress}
+      section={course.section}
+      url={url}
+    />
+  )
+}
+
+interface CourseDetailProps extends Omit<Course, 'activities'> {
+  url: string
+}
+
+const CourseCard = ({
+  courseDetails,
+  courseProgress,
+  section,
+  url,
+}: CourseDetailProps) => {
+  const theme = useMantineTheme()
+
+  const { sectionName, sectionSchedule, classMeetings } = section
+  const { courseName, courseCode } = courseDetails
+
+  const { currentMeeting } = useCurrentMeeting(classMeetings)
+  const [hovered, setHovered] = useState(false)
+
+  return (
+    <Card
+      component={Link}
+      to={url}
+      withBorder
+      radius="md"
+      p="xs"
+      shadow={hovered ? 'sm' : 'xs'}
+      style={{ cursor: 'pointer' }}
+      onMouseEnter={() => setHovered(true)}
+      onMouseLeave={() => setHovered(false)}
+    >
+      {/*Image*/}
+      <Flex pos="relative">
+        <Image
+          src="https://images.unsplash.com/photo-1511275539165-cc46b1ee89bf?q=80&w=1170&auto=format&fit=crop&ixlib=rb-4.1.0&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
+          alt="Norway"
+          radius="md"
+          miw="100%"
+          h="5.375rem"
+        />
+        <Text
+          pos="absolute"
+          top="4px"
+          left="4px"
+          size="0.625em"
+          fw={500}
+          c="white"
+          px="sm"
+          py="xs"
+          bdrs="md"
+          bg="black"
+          opacity={0.75}
+          style={{ textDecoration: hovered ? 'underline' : 'none' }}
+        >
+          {courseCode} • {sectionName}
+        </Text>
+      </Flex>
+      {/*Course Details*/}
+      <Group justify="space-between" wrap="nowrap" mt="xs">
+        <Stack gap={'md'} w={'16rem'} px={'xs'}>
+          <Stack mt={'xs'}>
+            <Title
+              order={3}
+              w={'100%'}
+              lineClamp={1}
+              c={theme.primaryColor}
+              style={{ textDecoration: hovered ? 'underline' : 'none' }}
+            >
+              {courseName}
+            </Title>
+            <Text
+              fw={400}
+              size={'sm'}
+              c={theme.colors.dark[3]}
+              style={{ textDecoration: hovered ? 'underline' : 'none' }}
+            >
+              {sectionSchedule.day} {sectionSchedule.time}
+            </Text>
+          </Stack>
+          <CourseCardActionButton
+            currentMeeting={currentMeeting}
+            courseCode={courseCode}
+          />
+          <Group justify="space-between">
+            <Group gap="0.25rem">
+              <Text fw={500} size={'xs'} c={theme.colors.dark[3]}>
+                Completed
+              </Text>
+              <Group gap="0">
+                <RingProgress
+                  size={20}
+                  thickness={3}
+                  sections={[
+                    {
+                      value: courseProgress * 100,
+                      color: theme.colors.blue[5],
+                    },
+                  ]}
+                />
+                <Text fw={500} size={'xs'} c={theme.colors.dark[3]}>
+                  {courseProgress * 100}%
+                </Text>
+              </Group>
+            </Group>
+            <CourseDashboardQuickActions />
+          </Group>
+        </Stack>
+      </Group>
+    </Card>
+  )
+}
+
+const CourseListRow = ({
+  courseDetails,
+  courseProgress,
+  section,
+  url,
+}: CourseDetailProps) => {
+  const theme = useMantineTheme()
+
+  const { sectionName, sectionSchedule, classMeetings } = section
+  const { courseName, courseCode } = courseDetails
+
+  const { currentMeeting } = useCurrentMeeting(classMeetings)
+  const [hovered, setHovered] = useState(false)
+
+  return (
+    <Card
+      component={Link}
+      to={url}
+      withBorder
+      radius="md"
+      p="0"
+      shadow={hovered ? 'sm' : 'xs'}
+      w={'100%'}
+      style={{ cursor: 'pointer' }}
+      onMouseEnter={() => setHovered(true)}
+      onMouseLeave={() => setHovered(false)}
+    >
+      <Group justify="space-between" wrap="nowrap">
+        <Box bg={theme.colors.primary[1]} h={'100%'} w={'20px'}></Box>
+        <Stack w={'65%'} p={'xs'} justify={'space-between'}>
+          <Group gap={'xs'}>
+            <Title
+              order={3}
+              lineClamp={1}
+              maw={'75%'}
+              c={theme.primaryColor}
+              style={{ textDecoration: hovered ? 'underline' : 'none' }}
+            >
+              {courseName}
+            </Title>
+            <CourseDashboardQuickActions />
+          </Group>
+          <Text
+            fw={400}
+            size={'sm'}
+            c={theme.colors.dark[3]}
+            style={{ textDecoration: hovered ? 'underline' : 'none' }}
+          >
+            {courseCode} • {sectionName} | {sectionSchedule.day}{' '}
+            {sectionSchedule.time}
+          </Text>
+        </Stack>
+        <Stack w={'30%'} p={'xs'} justify={'space-between'}>
+          <CourseCardActionButton
+            currentMeeting={currentMeeting}
+            courseCode={courseCode}
+          />
+          <Group gap="xs">
+            <Text fw={500} size={'xs'} c={theme.colors.dark[3]}>
+              Completed:
+            </Text>
+            <Progress color={theme.colors.blue[5]} value={50} w={'50%'} />
+            <Text fw={500} size={'xs'} c={theme.colors.dark[3]}>
+              {courseProgress * 100}%
+            </Text>
+          </Group>
+        </Stack>
+      </Group>
+    </Card>
+  )
+}
+
+type CourseCardActionButtonProps = {
+  currentMeeting?: ClassMeeting
+  courseCode: string
+}
+const CourseCardActionButton = ({
+  currentMeeting,
+  courseCode,
+}: CourseCardActionButtonProps) => {
+  const { authUser } = useAuth('protected')
+  return (
+    <Button
+      component={Link}
+      leftSection={
+        authUser.role === 'student' ? (
+          <IconVideo size={16} />
+        ) : (
+          <IconDeviceDesktop size={16} />
+        )
+      }
+      size="xs"
+      radius="xl"
+      variant="filled"
+      disabled={authUser.role === 'student' ? !currentMeeting : false}
+      to={
+        authUser.role === 'student'
+          ? currentMeeting?.meetingLink
+          : `/courses/${courseCode}/edit`
+      }
+    >
+      {authUser.role === 'student'
+        ? 'Join Meeting'
+        : authUser.role === 'mentor'
+          ? 'Start Meeting'
+          : 'Manage Content'}
+    </Button>
   )
 }
